requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.0
python-dotenv==0.20.0
Werkzeug==2.1.2
zipp==3.8.0
